+++ tools/names.sh
@@ -75,5 +75,5 @@ function get_MP_D_CONTAINER_x {
   local profile="${2}"
   local name="${3}"
 
-  echo "$(get_MP_D_PROFILE_x "${config_file}" "${profile}")_${name}_1"
+  echo "$(get_MP_D_PROFILE_x "${config_file}" "${profile}")-${name}-1"
 }